// types/response/dormitories.ts
export type DormitoriesResponseType = {
    id: number;
    work_place_id: number;
    note: string;
    name_en: string;
    name_zh: string;
    name_vn: string;
    available: null;
    code: string;
    updated_by: string;
    capacity: number;
    created_by: string;
    created_at: string;
    updated_at: string;
    active: boolean;
};